SM2: Add distid parameter for signature sign and verify
According to the specifications GM/T 0003.2-2012 and GM/T 0003.5-2012,
the SM2 algorithm signature requires a Za value when generating the hash
required for signature. The Za value is generated by the fixed parameter
of the elliptic curve, the public key and the user entered identifiable ID.
This patch adds a distid parameter to specify the ID during signature sign
and verification.

An example usage for generating a self signed CA certificate is as follows:

  openssl ecparam -genkey -name SM2 -text -out ca.key
  openssl req -verbose -new -days 10000 -x509 \
      -sm3 -sigopt "distid:1234567812345678" \
      -config genkey.conf \
      -key ca.key \
      -out ca.crt

The genkey.conf file content is as follows:

  [ req ]
  distinguished_name = req_distinguished_name
  prompt = no
  string_mask = utf8only
  x509_extensions = v3_ca

  [ req_distinguished_name ]
  O = rfc8998-CA
  CN = RFC8998 certificate signing key
  emailAddress = ca@rfc8998-ca

  [ v3_ca ]
  basicConstraints=CA:TRUE
  subjectKeyIdentifier=hash
  authorityKeyIdentifier=keyid:always,issuer

  [ skid ]
  basicConstraints=CA:TRUE
  subjectKeyIdentifier=12345678
  authorityKeyIdentifier=keyid:always,issuer

Signed-off-by: Tianjia Zhang <tianjia.zhang@linux.alibaba.com>